requirements.txt
aiohttp==3.9.5
brotli==1.1.0
ConfigParser==7.0.0
cryptography==42.0.5
Cython==3.0.10
discord.py==2.3.2
dl==0.1.0
docutils==0.21.2
HTMLParser==0.0.2
importlib_metadata==7.1.0
ipaddr==2.2.0
keyring==25.2.0
lockfile==0.12.2
lxml==5.2.1
mypy_extensions==1.0.0
numpy==1.26.4
ordereddict==1.1
protobuf==5.26.1
psycopg==3.1.18
psycopg_pool==3.2.1
pyOpenSSL==24.1.0
python-dotenv==1.0.1
pywin32==306
thread==2.0.1
typing_extensions==4.11.0
wincertstore==0.2.1
xmlrpclib==1.0.1
zipp==3.18.1
